add test for simple_example / simple_compressed_example

+++ src/nodelet/simple_example_nodelet.cpp
@@ -59,6 +59,7 @@ class ImageConverter
   image_transport::ImageTransport it_;
   image_transport::Subscriber image_sub_;
   image_transport::Publisher image_pub_;
+  bool debug_view_;
 
 public:
   ImageConverter()
@@ -69,12 +70,18 @@ public:
       &ImageConverter::imageCb, this);
     image_pub_ = it_.advertise("/image_converter/output_video/raw", 1);
 
-    cv::namedWindow(OPENCV_WINDOW);
+    ros::NodeHandle pnh_("~");
+    pnh_.param("debug_view", debug_view_, false);
+    if( debug_view_) {
+      cv::namedWindow(OPENCV_WINDOW);
+    }
   }
 
   ~ImageConverter()
   {
-    cv::destroyWindow(OPENCV_WINDOW);
+    if( debug_view_) {
+      cv::destroyWindow(OPENCV_WINDOW);
+    }
   }
 
   void imageCb(const sensor_msgs::ImageConstPtr& msg)
@@ -94,9 +101,11 @@ public:
     if (cv_ptr->image.rows > 110 && cv_ptr->image.cols > 110)
       cv::circle(cv_ptr->image, cv::Point(cv_ptr->image.cols/2, cv_ptr->image.rows/2), 100, CV_RGB(255,0,0));
 
-    // Update GUI Window
-    cv::imshow(OPENCV_WINDOW, cv_ptr->image);
-    cv::waitKey(3);
+    if( debug_view_) {
+      // Update GUI Window
+      cv::imshow(OPENCV_WINDOW, cv_ptr->image);
+      cv::waitKey(3);
+    }
 
     // Output modified video stream
     image_pub_.publish(cv_ptr->toImageMsg());
